requirements.txt
dill
pandas
numpy
scikit-learn==1.4.0
dataclasses
requests
datetime
python-dotenv
streamlit
#-e .